+++ crates/symbolicator/src/services/download/http.rs
@@ -8,7 +8,7 @@ use std::time::Duration;
 
 use anyhow::Result;
 use futures::prelude::*;
-use reqwest::{header, Client};
+use reqwest::{header, Client, StatusCode};
 use url::Url;
 
 use super::{
@@ -112,6 +112,12 @@ impl HttpDownloader {
                     let stream = response.bytes_stream().map_err(DownloadError::Reqwest);
 
                     super::download_stream(source, stream, destination, timeout).await
+                } else if matches!(
+                    response.status(),
+                    StatusCode::FORBIDDEN | StatusCode::UNAUTHORIZED
+                ) {
+                    log::debug!("Insufficient permissions to download from {}", download_url);
+                    Err(DownloadError::Permissions)
                 // If it's a client error, chances are either it's a 404 or it's permission-related.
                 } else if response.status().is_client_error() {
                     log::debug!(