+++ crates/symbolicator/src/services/download/mod.rs
@@ -15,6 +15,7 @@ use thiserror::Error;
 use tokio::fs::File;
 use tokio::io::AsyncWriteExt;
 
+use crate::cache::CacheStatus;
 use crate::utils::futures::{self as future_utils, m, measure};
 use crate::utils::paths::get_directory_paths;
 
@@ -52,6 +53,8 @@ pub enum DownloadError {
     Sentry(#[from] sentry::SentryError),
     #[error("failed to fetch data from S3")]
     S3(#[from] s3::S3Error),
+    #[error("missing permissions for file")]
+    Permissions,
     /// Typically means the initial HEAD request received a non-200, non-400 response.
     #[error("failed to download: {0}")]
     Rejected(StatusCode),
@@ -60,13 +63,34 @@ pub enum DownloadError {
 }
 
 impl DownloadError {
+    /// This produces a user-facing string representation of a download error if it is a variant
+    /// that needs to be stored as a [`CacheStatus::CacheSpecificError`] entry in the download cache.
     pub fn for_cache(&self) -> String {
         match self {
             DownloadError::Gcs(inner) => format!("{}: {}", self, inner),
             DownloadError::Sentry(inner) => format!("{}: {}", self, inner),
             DownloadError::S3(inner) => format!("{}: {}", self, inner),
+            DownloadError::Permissions => self.to_string(),
             DownloadError::CachedError(original_message) => original_message.clone(),
             _ => format!("{}", self),
+        }
+    }
+
+    /// If a given cache entry is [`CacheStatus::CacheSpecificError`], this parses and extracts its
+    /// contents into a [`DownloadError`]. This will return none if a
+    /// non-[`CacheStatus::CacheSpecificError`] is provided.
+    pub fn from_cache(status: &CacheStatus) -> Option<Self> {
+        match status {
+            CacheStatus::Positive => None,
+            CacheStatus::Negative => None,
+            CacheStatus::Malformed(_) => None,
+            CacheStatus::CacheSpecificError(message) => {
+                if message.starts_with(&Self::Permissions.to_string()) {
+                    Some(Self::Permissions)
+                } else {
+                    Some(Self::CachedError(message.clone()))
+                }
+            }
         }
     }
 }