+++ crates/symbolicator/src/services/download/sentry.rs
@@ -283,7 +283,12 @@ impl SentryDownloader {
                     let stream = response.bytes_stream().map_err(DownloadError::Reqwest);
 
                     super::download_stream(source, stream, destination, timeout).await
-                // If it's a client error, chances are either it's a 404 or it's permission-related.
+                } else if matches!(
+                    response.status(),
+                    StatusCode::FORBIDDEN | StatusCode::UNAUTHORIZED
+                ) {
+                    log::debug!("Insufficient permissions to download from {}", download_url);
+                    Err(DownloadError::Permissions)
                 } else if response.status().is_client_error() {
                     log::debug!(
                         "Unexpected client error status code from {}: {}",